+++ src/KYLib/MathFn/Math.cs
@@ -1,5 +1,5 @@
 using System;
-using KYLib.Interfaces;
+using KYLib.Abstractions;
 namespace KYLib.MathFn;
 
 // Aqui se guardan los metodos que proveen implementacion de los metodos de Math.