+++ src/KYLib/MathFn/Mathf.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using KYLib.Interfaces;
+using KYLib.Abstractions;
 using KYLib.Utils;
 namespace KYLib.MathFn;