+++ src/KYLib/MathFn/Real20.cs
@@ -1,8 +1,8 @@
 ﻿#if NETSTANDARD2_0
 using System;
 using System.Globalization;
+using KYLib.Abstractions;
 using KYLib.Helpers;
-using KYLib.Interfaces;
 namespace KYLib.MathFn
 {
 	partial struct Real : INumber<double>