+++ src/KYLib/MathFn/Small.cs
@@ -1,6 +1,6 @@
 using System;
+using KYLib.Abstractions;
 using KYLib.Helpers;
-using KYLib.Interfaces;
 namespace KYLib.MathFn;
 
 /// <summary>