+++ src/KYLib/Modding/Mod.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Reflection;
-using KYLib.Interfaces;
+using KYLib.Abstractions;
 using KYLib.Utils;
 namespace KYLib.Modding;